Rename html_tag_collector to source_tag_collector
