+++ identification_pipeline.py
@@ -2,7 +2,7 @@ import requests
 import os
 import polars as pl
 import sys
-from html_tag_collector.collector import collector_main
+from source_text_collector.collector import collector_main
 from agency_identifier.identifier import match_urls_to_agencies_and_clean_data
 from datetime import datetime as dt
 from dotenv import load_dotenv
