+++ source_text_collector/collector.py
